increase IT execute speed, fix commandline bug
